pyproject.toml
[project]
name = "eule"
version = "0.1.0"
description = "Trade Journal + Portfolio Monitor CLI"
authors = [
    {name = "blaurock"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "typer>=0.15.0",
    "psycopg[binary]>=3.2.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "loguru>=0.7.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0",
    "ibind[oauth]>=0.1.21",
    "trading-ig>=0.0.23",
    "yfinance>=0.2.55",
]

[project.scripts]
eule = "eule.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
